% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhRunModel.R
\name{checkOperatingSystem}
\alias{checkOperatingSystem}
\title{Helper function: checks whether operating system is windows or linux}
\usage{
checkOperatingSystem()
}
\description{
Helper function: checks whether operating system is windows or linux
}
\examples{
os <- checkOperatingSystem() ### if function shell() is not available -> linux"
os ### return operating system
}
