% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{convBasic}
\alias{convBasic}
\title{Helper function: convert basic config to  FORTRAN style}
\usage{
convBasic(baseConf, dbg = TRUE)
}
\arguments{
\item{baseConf}{with basic parameterisation, i.e. conf$basic}

\item{dbg}{If TRUE debug messages are printed on the screen (default: TRUE)}
}
\value{
Basic configuration in FORTRAN style
}
\description{
Helper function: convert basic config to  FORTRAN style
}
