% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{convertStringVectorToMatrix}
\alias{convertStringVectorToMatrix}
\title{Helper function: convert string vector to numeric matrix}
\usage{
convertStringVectorToMatrix(values, splitSep = " ", byrow = TRUE,
  rows = length(values), toNumeric = TRUE)
}
\arguments{
\item{values}{character vector to be converted to matrix}

\item{splitSep}{seperator used as split parameter in function strsplit (Default: " ")}

\item{byrow}{If TRUE matrix is filled by rows first. Default: TRUE}

\item{rows}{for constructing matrix: Default: length(values)}

\item{toNumeric}{If true values will be converted to numeric, if FALSE not}
}
\value{
Returns matrix with nrows equals to length of values input vector
}
\description{
Helper function: convert string vector to numeric matrix
}
