% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhRunModel.R
\name{convertToWindowsPath}
\alias{convertToWindowsPath}
\title{Helper function: converts string to windows path}
\usage{
convertToWindowsPath(path, sep = "\\\\/")
}
\arguments{
\item{path}{path string to be converted to windows path}

\item{sep}{seperator which default: "/", which is used on unix and for R)}
}
\description{
Helper function: converts string to windows path
}
\examples{
unixPath <- "/usr/model/vs2dh"
winPath <- convertToWindowsPath(path=unixPath)
winPath
}

