% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{convInitial}
\alias{convInitial}
\title{Helper function: converting initial conditions to FORTRAN style}
\usage{
convInitial(iniConf, dbg = TRUE)
}
\arguments{
\item{iniConf}{config with soils parameterisation, i.e. conf$initial}

\item{dbg}{If TRUE debug messages are printed on the screen (default: TRUE)}
}
\value{
Soil configuration in FORTRAN style
}
\description{
Helper function: converting initial conditions to FORTRAN style
}

