% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{convMatrixByRowToString}
\alias{convMatrixByRowToString}
\title{Helper function: converts matrix to one string}
\usage{
convMatrixByRowToString(matr, colSep = " ", rowSep = "\\n")
}
\arguments{
\item{matr}{matrix with character data}

\item{colSep}{seperator used for merging data from all columns for each row
(default: " ")}

\item{rowSep}{seperator used for separating the data of different rows
(default: "newline")}
}
\value{
One string with each row separated with rowSep
}
\description{
Helper function: converts matrix to one string
}

