% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{convRecharge}
\alias{convRecharge}
\title{Helper function: converting recharge periods to FORTRAN style}
\usage{
convRecharge(rechConf, dbg = TRUE)
}
\arguments{
\item{rechConf}{config with recharge period parameterisation, i.e. conf$recharge}

\item{dbg}{If TRUE debug messages are printed on the screen (default: TRUE)}
}
\value{
Recharge periods configuration in FORTRAN style
}
\description{
Helper function: converting recharge periods to FORTRAN style
}
