% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{convSoils}
\alias{convSoils}
\title{Helper function: converting soils structure to FORTRAN style}
\usage{
convSoils(conf, dbg = TRUE)
}
\arguments{
\item{conf}{config with R model parameterisation, i.e. conf}

\item{dbg}{If TRUE debug messages are printed on the screen (default: TRUE)}
}
\value{
Soil configuration in FORTRAN style
}
\description{
Helper function: converting soils structure to FORTRAN style
}

