% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{cutStringByPattern}
\alias{cutStringByPattern}
\title{Helper function: Cut string by pattern}
\usage{
cutStringByPattern(pattern, endString = TRUE, txt)
}
\arguments{
\item{pattern}{pattern to be searched for in string  \code{txt}; if pattern 
is a vector of two the string is cut between pattern[1] and pattern[2]}

\item{endString}{if TRUE the string after the search pattern is returned, if 
FALSE from the beginning until the search pattern (only evaluated if pattern 
\code{pattern} is vector of length 1!)}

\item{txt}{string or vector of strings to be searched for in string}
}
\value{
Returns only partial string(s) depending on the cut pattern and until 
the end of each string
}
\description{
Helper function: Cut string by pattern
}
\examples{
### Path to your vs2dh model directory
string = "A B C D"
pattern = "B "
cutStringByPattern(pattern=pattern, txt=string)
}
