% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{filterLines}
\alias{filterLines}
\title{Helper function: filtering lines}
\usage{
filterLines(pattern, patternNameCol = "txt", data, startLineOffset = 0,
  input, lineCol = "Line")
}
\arguments{
\item{pattern}{pattern which lines are searched}

\item{patternNameCol}{for pattern like "/B25" use: "txt", for para names 
use "paraNames", (Default: "txt")}

\item{data}{as retrieved by prepareImport()$res}

\item{startLineOffset}{offset of starting line: 0 (if parName is in same row 
as parVal, 1 if parVal is one row after parName)}

\item{input}{as retrieved by prepareImport()$input}

\item{lineCol}{Default "Line"}
}
\description{
Helper function: filtering lines
}
