% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{fortranFormat}
\alias{fortranFormat}
\title{Helper function: formatting R elments to FORTRAN style}
\usage{
fortranFormat(conf)
}
\arguments{
\item{conf}{as retrieved by vs2dhConfigure()}
}
\value{
config with prepared FORTRAN formatting style (only TRUE/FALSE values
will be replaced during vs2dh.writeConfig()
}
\description{
Helper function: formatting R elments to FORTRAN style
}
