% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{importMatrices}
\alias{importMatrices}
\title{Helper function: imports matrices (jtex, hvalues, tvalues)}
\usage{
importMatrices(prepData, dbg = TRUE)
}
\arguments{
\item{prepData}{object as retrieved by prepareImport()}

\item{dbg}{prints debug information on the screen}
}
\value{
Matrix values (always: soil, if available: initial pressure head &
temperature distribution)
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
inp <- prepareImport(model.path)
grid <- importMatrices(inp) 
#### Soil properties matrix
vs2dh.plotMatrix(data = grid$jtex)
#### Initial temperature distribution matrix
vs2dh.plotMatrix(data = grid$tvalues)
}
\description{
Helper function: imports matrices (jtex, hvalues, tvalues)
}
