% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{importMatrix}
\alias{importMatrix}
\title{Helper function: imports matrix values of vs2dh.dat}
\usage{
importMatrix(prepData, pattern = "JTEX", patternNameCol = "parNames",
  startLineOffset = 0, msg = "B9: Importing soil matrix...",
  nrow = NULL, dbg = TRUE)
}
\arguments{
\item{prepData}{object as retrieved by prepareImport()}

\item{pattern}{as used in filterLines()}

\item{patternNameCol}{as used in filterLines()}

\item{startLineOffset}{as used in filterLines()}

\item{msg}{message which parameter is imported. used if DBG == TRUE}

\item{nrow}{number of rows of matrix: Default: nly}

\item{dbg}{If TRUE prints debug information on the screen (Default: TRUE)}
}
\value{
Matrix values (always: soil, if available: initial pressure head &
temperature distribution)
}
\description{
Helper function: imports matrix values of vs2dh.dat
}
