% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{importRechargePeriods}
\alias{importRechargePeriods}
\title{Helper function: imports recharge periods}
\usage{
importRechargePeriods(prepData)
}
\arguments{
\item{prepData}{object as retrieved by prepareImport()}
}
\value{
Recharge periods (i.e. PART C) with naming parameter names in R style
}
\description{
xxxxxx
}
