% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{importSingleLineParas}
\alias{importSingleLineParas}
\title{Helper function: imports single line parameters}
\usage{
importSingleLineParas(prepData, dbg = TRUE)
}
\arguments{
\item{prepData}{object as retrieved by prepareImport()}

\item{dbg}{prints debug information on the screen}
}
\value{
Single line values with naming parameter names in R style (before 
part C, i.e. recharge period 1 starts)
}
\description{
FORTRAN parameters with always correspond to only one value
}
