% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{labeledList}
\alias{labeledList}
\title{Helper function: replace a list index with a name}
\usage{
labeledList(inpLst, labelTxt = "id")
}
\arguments{
\item{inpLst}{input list}

\item{labelTxt}{user defined txt used for naming list}
}
\value{
Labeled list
}
\description{
Helper function: replace a list index with a name
}

