% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConvertToFortran.R
\name{leafValues}
\alias{leafValues}
\title{Helper function: names of all sublists of a list returns the names of all
sublists of \emph{x} in the "$"-notation, e.g.
list$sublist$subsublist$subsubsublist}
\usage{
leafValues(x, values = list())
}
\arguments{
\item{x}{R list.}

\item{values}{name to be used as prefix for all names found. default: ""}
}
\value{
leafs of sublists into one dimensional list
}
\description{
Helper function: names of all sublists of a list returns the names of all
sublists of \emph{x} in the "$"-notation, e.g.
list$sublist$subsublist$subsubsublist
}
