% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{multipleLineValues}
\alias{multipleLineValues}
\title{Helper function: multipleLineValues}
\usage{
multipleLineValues(parName = "DELZ", data, patternNameCol = "parNames",
  startLineOffset = 1, toNumeric = TRUE)
}
\arguments{
\item{parName}{e.g. "PLTIM", "DELZ" or "DXR"}

\item{data}{as retrieved by  prepareImport()}

\item{patternNameCol}{for pattern like "/B25" use: "txt", for para names 
use "paraNames", (Default: "txt") -> used by function: filterLines()}

\item{startLineOffset}{offset of starting line: 0 (if parName is in same row 
as parVal, 1 if parVal is one row after parName), -> used by function: filterLines()}

\item{toNumeric}{If TRUE result is converted to numeric (Default: TRUE)}
}
\value{
List with one vector of multiple lines values for each parameter
}
\description{
Writes one vector for values in multiple lines  
that beginn in FORTRAN "in the next line" (e.g. DELZ, DXR)
}
