% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{nodePairs}
\alias{nodePairs}
\title{Helper function: node pairs}
\usage{
nodePairs(j, n)
}
\arguments{
\item{j}{Row of each cell}

\item{n}{Column of each cell}
}
\value{
data.frame with node pairs
}
\description{
Helper function: node pairs
}
