% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhReadOutput.R
\name{patternSelect}
\alias{patternSelect}
\title{Helper function: selects pattern contained in different lines of a string vector
(e.g. imported using readLines())}
\usage{
patternSelect(pattern, data)
}
\arguments{
\item{pattern}{index of row to be split (valid values: 1-3)}

\item{data}{vector with strings}
}
\value{
data.frame with columns lines (indicating the lines where pattern was
found and txt (string contained in that lines)
}
\description{
Helper function: selects pattern contained in different lines of a string vector
(e.g. imported using readLines())
}

