% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadOutput.R
\name{splitHeader}
\alias{splitHeader}
\title{Helper function: splits header of output file with (energy, fluid) mass 
balance time series (balance.out)}
\usage{
splitHeader(row, header)
}
\arguments{
\item{row}{index of row to be split (valid values: 1-3)}

\item{header}{first three lines of balance.out containing header}
}
\value{
splitted header
}
\description{
Helper function: splits header of output file with (energy, fluid) mass 
balance time series (balance.out)
}
