% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.Configure}
\alias{vs2dh.Configure}
\title{Configure complete vs2dh parameterisation}
\usage{
vs2dh.Configure(basic = vs2dh.ConfigureBasic(),
  soils = vs2dh.ConfigureSoils(), initial = vs2dh.ConfigureInitial(),
  recharge = vs2dh.ConfigureRechargePeriods())
}
\arguments{
\item{basic}{parameterisation (title, units, time, output, solver, grid, options), 
as retrieved by  vs2dh.ConfigureBasic(),}

\item{soils}{retrieved by vs2dh.ConfigureSoils()}

\item{initial}{initial conditions for flow (pressure head or moisture contents and transport 
(temperatures), as retrieved by  vs2dh.ConfigureInitial()}

\item{recharge}{recharge periods as retrieved by vs2dh.ConfigureRechargePeriods()}
}
\value{
Complete vs2dh parameterisation in R style
}
\description{
Configure complete vs2dh parameterisation
}
