% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureBasic}
\alias{vs2dh.ConfigureBasic}
\title{Configure basic model parameters}
\usage{
vs2dh.ConfigureBasic(titl = "My title", units = vs2dh.ConfigureBasicUnits(),
  time = vs2dh.ConfigureBasicTime(), output = vs2dh.ConfigureBasicOutput(),
  solver = vs2dh.ConfigureBasicSolver(), grid = vs2dh.ConfigureBasicGrid(),
  options = vs2dh.ConfigureBasicOptions())
}
\arguments{
\item{titl}{80-character problem description (formatted read, 20A4)}

\item{units}{as retrieved by  vs2dh.ConfigureBasicUnits()}

\item{time}{as retrieved by  vs2dh.ConfigureBasicTime()}

\item{output}{as retrieved by vs2dh.ConfigureBasicTime()}

\item{solver}{retrieved by vs2dh.ConfigureBasicSolver()}

\item{grid}{retrieved by vs2dh.ConfigureBasicGrid()}

\item{options}{as simulation options as retrieved by vs2dh.ConfigureBasicOptions
(default: TRUE)}
}
\value{
Basic configuration
}
\description{
Configure basic model parameters
}
