% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureBasicOptions}
\alias{vs2dh.ConfigureBasicOptions}
\title{Configure basic simulation options: energy transport, evaporation or 
evapotranspiration should be simulated? (FUNCTIONALITY NOT IMPLEMENTED YET! 
DO NOT CHANGE default PARAMETERISATION)}
\usage{
vs2dh.ConfigureBasicOptions(trans = TRUE, bcit = FALSE, etsim = FALSE)
}
\arguments{
\item{trans}{Should energy transport be simulated? (default: TRUE), Option
FALSE currently NOT IMPLEMENTED YET!!!!!!!}

\item{bcit}{NOT IMPLEMENTED YET!!!!!!! If TRUE evaporation is to be simulated 
for this simulation;if FALSE no evaporation is simulated for this simulation
(default: FALSE)}

\item{etsim}{NOT IMPLEMENTED YET!!!!!!! If TRUE evapotanspiration (plant-root 
extraction) is to be simulated for this simulation; if FALSE no evaporation is simulated for 
this simulation (default: FALSE).}
}
\value{
Return basic simulation options
}
\description{
Configure basic simulation options: energy transport, evaporation or 
evapotranspiration should be simulated? (FUNCTIONALITY NOT IMPLEMENTED YET! 
DO NOT CHANGE default PARAMETERISATION)
}
