% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureBasicOutputMain}
\alias{vs2dh.ConfigureBasicOutputMain}
\title{Configure main output file ("vs2dh.out")}
\usage{
vs2dh.ConfigureBasicOutputMain(f6p = TRUE, thpt = TRUE, spnt = TRUE,
  ppnt = TRUE, hpnt = TRUE, vpnt = TRUE)
}
\arguments{
\item{f6p}{If TRUE  mass balance is to be written to file "vs2dt.out" 
for each time step; if FALSE mass balance is to be written to file "vs2dt.out" 
only at observation times and ends of recharge periods (default: TRUE)}

\item{thpt}{If TRUE volumetric moisture contents are to be written to 
file "vs2dt.out", otherwise: FALSE (default: TRUE)}

\item{spnt}{If TRUE saturation is written to file "vs2dt.out", 
otherwise: FALSE (default: TRUE)}

\item{ppnt}{If TRUE pressure heads are written to file "vs2dt.out", 
otherwise: FALSE (default: TRUE)}

\item{hpnt}{If TRUE total heads are written to file "vs2dt.out", 
otherwise: FALSE (default: TRUE)}

\item{vpnt}{If TRUE velocities are written to file "vs2dt.out", 
otherwise: FALSE (default: TRUE)}
}
\value{
Output config of vs2dh.out file
}
\description{
Configure main output file ("vs2dh.out")
}
