% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureBasicOutputTimes}
\alias{vs2dh.ConfigureBasicOutputTimes}
\title{Configure model output times}
\usage{
vs2dh.ConfigureBasicOutputTimes(pltim = seq(0, 10, 0.5), numt = 1000,
  enhancedPrecision = FALSE, dbg = TRUE)
}
\arguments{
\item{pltim}{Elapsed times at which pressure heads and temperatures are to be
written to file "variables.out", and heads, temperatures, saturations, 
velocities, and/or moisture contents to file "vs2dt.out", T. (default: seq(0,10,0.5))}

\item{numt}{Maximum number of time steps. (default: 1000), will be set 
automatically if called from higher level function vs2dh.ConfigureBasic()}

\item{enhancedPrecision}{if TRUE enhanced results are written with enhanced 
precision to files6 (vs2dh.out) & file9 (variables.out), (default: FALSE)}

\item{dbg}{if TRUE additional information printed on screen (default: TRUE)}
}
\value{
Model output times
}
\description{
Configure model output times
}
