% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureBasicTime}
\alias{vs2dh.ConfigureBasicTime}
\title{Configure time parameters}
\usage{
vs2dh.ConfigureBasicTime(stim = 0, tmax = 10.1)
}
\arguments{
\item{stim}{Initial time (usually set to 0). (default: 0)}

\item{tmax}{Maximum simulation time. (default: 10.1)}
}
\value{
Time parameters
}
\description{
Configure time parameters
}

