% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureBasicUnits}
\alias{vs2dh.ConfigureBasicUnits}
\title{Configure unit parameters}
\usage{
vs2dh.ConfigureBasicUnits(zunit = "m", tunit = "day", cunx = "J")
}
\arguments{
\item{zunit}{Units used for length, "m" for meters. (default: "m")}

\item{tunit}{Units used for time, "sec" for seconds, "day" for day 
(default: "day")}

\item{cunx}{Units used for heat, "J" for Joules. (default: "J")}
}
\value{
Model units
}
\description{
Configure unit parameters
}
