% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureGenuchten}
\alias{vs2dh.ConfigureGenuchten}
\title{Configure Genuchten flow parameters}
\usage{
vs2dh.ConfigureGenuchten(ratioKzKh = 1, satKh = 750, ss = 0,
  porosity = 0.39, alpha = 2.3, rmc = 0.05, beta = 5.8)
}
\arguments{
\item{ratioKzKh}{Ratio of hydraulic conductivity in the z-coordinate direction
to that in the x-coordinate direction (Default:1)}

\item{satKh}{Saturated hydraulic conductivity (K) at 20 C in the x-coordinate
direction for class ITEX, L/T. (Default: 750 m/d)}

\item{ss}{Specific storage (Ss), L^-1. (Default: 0)}

\item{porosity}{Porosity (f), (Default: 0.39)}

\item{alpha}{van Genuchten alpha. NOTE: alpha is as defined by van Genuchten
(1980) and is the negative reciprocal of alpha' used in earlier versions (prior
to version 3.0) of VS2DT, L., (Default: 2.3/m)}

\item{rmc}{Residual moisture content,  (Default: 0.05)}

\item{beta}{van Genuchten parameter, beta' in Healy (1990) and Lappala and
others (1987), (Default:5.8)}
}
\value{
Genuchten flow parameters
}
\description{
Configure Genuchten flow parameters
}
\seealso{
\url{http://pubs.usgs.gov/circ/2003/circ1260/pdf/Circ1260.pdf} p.87 for default parameterisation
}

