% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureInitial}
\alias{vs2dh.ConfigureInitial}
\title{Configure initial conditions}
\usage{
vs2dh.ConfigureInitial(flow = vs2dh.ConfigureInitialFlow(),
  temp = vs2dh.ConfigureInitialTemp())
}
\arguments{
\item{flow}{initial flow conditions (i.e. pressure head or saturation distribution
for model domains, as retrieved by vs2dh.ConfigureInitialFlow()}

\item{temp}{initial temperature distribution as retrieved by
vs2dh.ConfigureInitialTemp()}
}
\value{
Initial model conditions
}
\description{
Configure initial conditions
}
