% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureInitialTemp}
\alias{vs2dh.ConfigureInitialTemp}
\title{Configure initial temperature distribution:}
\usage{
vs2dh.ConfigureInitialTemp(values = 10)
}
\arguments{
\item{values}{either constant (for constant initial temperature for whole model
domain, or matrix with number of rows (equals to nxr) and columns (equals to nly),
with user defined initial temperature distribution (default: 10 degree C,
constant temperature)}
}
\value{
Initial temperature distribution parameterisation
}
\description{
Configure initial temperature distribution:
}

