% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureRechargePeriod}
\alias{vs2dh.ConfigureRechargePeriod}
\title{Configure recharge period}
\usage{
vs2dh.ConfigureRechargePeriod(tper = 0.1,
  options = vs2dh.ConfigureRechargePeriodOptions(),
  solver = vs2dh.ConfigureRechargePeriodSolver(),
  boundary = vs2dh.ConfigureBoundaryCondition())
}
\arguments{
\item{tper}{Length of this recharge period, T. (default: 0.1)}

\item{options}{as retrieved by vs2dh.ConfigureRechargePeriodOptions()}

\item{solver}{as retrieved by vs2dh.ConfigureRechargePeriodSolver()}

\item{boundary}{as retrieved by vs2dh.ConfigureBoundaryCondition()}
}
\value{
Recharge period
}
\description{
Configure recharge period
}

