% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureRechargePeriods}
\alias{vs2dh.ConfigureRechargePeriods}
\title{Configure recharge periods}
\usage{

  vs2dh.ConfigureRechargePeriods(periods = list(vs2dh.ConfigureRechargePeriod(),
  vs2dh.ConfigureRechargePeriod(tper = 10)))
}
\arguments{
\item{periods}{list with one or multiple recharge periods with structure as retrieved by
vs2dh.ConfigureRechargePeriod()}
}
\value{
List with parameterisation of recharge periods
}
\description{
Configure recharge periods
}

