% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureSeepage}
\alias{vs2dh.ConfigureSeepage}
\title{Configure seepage}
\usage{
vs2dh.ConfigureSeepage(seepFaces = list(vs2dh.ConfigureSeepageFace(),
  vs2dh.ConfigureSeepageFace(j = 30:20, n = 45)))
}
\arguments{
\item{seepFaces}{list of seepage faces as retrieved by vs2dh.ConfigureSeepageFace()}
}
\value{
Seepage config
}
\description{
Configure seepage
}
