% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureSoil}
\alias{vs2dh.ConfigureSoil}
\title{Configure soil parameters}
\usage{
vs2dh.ConfigureSoil(hk = vs2dh.ConfigureGenuchten(),
  ht = vs2dh.ConfigureTrans())
}
\arguments{
\item{hk}{flow properties \code{vs2dh.ConfigureGenuchten()}}

\item{ht}{transport properties \code{vs2dh.ConfigureTrans()}}
}
\value{
(Genuchten) flow & transport parameters of soil
}
\description{
Configure soil parameters
}
