% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureSoilGrid}
\alias{vs2dh.ConfigureSoilGrid}
\title{Configure soil grid}
\usage{
vs2dh.ConfigureSoilGrid(ntex = 2, grid = vs2dh.ConfigureBasicGrid(),
  irow = 0)
}
\arguments{
\item{ntex}{number of different textual classes (default: 2)}

\item{grid}{grid matrix as retrieved by \code{vs2dh.ConfigureBasicGrid()}}

\item{irow}{Textural classes are read for each row (irow=0). This option is preferable if many}
}
\value{
Grid with soil indexes
}
\description{
Configure soil grid
}
