% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureSoils}
\alias{vs2dh.ConfigureSoils}
\title{Configure Soils}
\usage{
vs2dh.ConfigureSoils(props = list(vs2dh.ConfigureSoil(),
  vs2dh.ConfigureSoil(hk = vs2dh.ConfigureGenuchten(ratioKzKh = 100))),
  grid = vs2dh.ConfigureSoilGrid(ntex = length(props)), wus = 0.5)
}
\arguments{
\item{props}{Soil (flow & transport) properties as retrieved by \code{vs2dh.ConfigureSoil()} 
ordered by itex number, i.e. first list element = itex1, second = itex2}

\item{grid}{grid matrix as retrieved by \code{vs2dh.ConfigureSoilGrid}}

\item{wus}{Weighting option for intercell relative hydraulic conductivity: 1 
(for full upstream weighting), 0.5 (for arithmetic mean) and 0.0 (for geometric 
mean), (default: 0.5)}
}
\value{
Configuration of soils for flow and energy transport (only hydraulic
function type 1 vanGenuchten model is implemented!)
}
\description{
Configure Soils
}
