% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhConfigure.R
\name{vs2dh.ConfigureTrans}
\alias{vs2dh.ConfigureTrans}
\title{Configure soil transport parameters}
\usage{
vs2dh.ConfigureTrans(alphaL = 1, alphaT = 0.1, cs = 2180000,
  ktRmc = 129600, ktSat = 155520, cw = 4180000)
}
\arguments{
\item{alphaL}{Longitudinal dispersivity, L. (Default: 1 m)}

\item{alphaT}{Transverse dispersivity, L. (Default: 0.1 m)}

\item{cs}{Heat capacity of dry solids (Cs), Q/L3 C. (Default: 2180000.0
J/m3C)}

\item{ktRmc}{Thermal conductivity of water sediment at residual moisture
content, Q/LTC. (default: 129600.0)}

\item{ktSat}{Thermal conductivity of water sediments at full saturation,
Q/LC. (Default: 155520.0)}

\item{cw}{Heat capacity of water (Cw), which is the product of density times
specific heat of water, Q/L3 C. (default: 4180000.0)}
}
\value{
Soil transport parameters
}
\description{
Configure soil transport parameters
}
\seealso{
\url{http://pubs.usgs.gov/circ/2003/circ1260/pdf/Circ1260.pdf} p.87
  for default parameterisation
}
