% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhPlot.R
\name{vs2dh.plotMassBalance}
\alias{vs2dh.plotMassBalance}
\title{Plotting of mass balance time series}
\usage{
vs2dh.plotMassBalance(paras = c("TOTAL__FLOW_IN", "TOTAL__FLOW_OUT",
  "FLUID__STORAGE"), paraUnit = "TIMESTEP", data,
  mainLabel = "Flow mass balance", ...)
}
\arguments{
\item{paras}{vector with at least one or multiple parameters in model results, 
for checking available paras run (see example: validParas). The parameter "TIME"
is not allowed!}

\item{paraUnit}{TIMESTEP RATE or}

\item{data}{as returned by vs2di.run()$balance}

\item{mainLabel}{a text to be written above the plot}

\item{...}{further parameters passed to \code{xyplot()}}
}
\value{
Temporal time series plot of mass balance variables
}
\description{
Plotting of mass balance time series
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)  
#### Checking available parameter names for "paras":
validParas <- gsub(pattern="__TOTAL|__RATE|__TIMESTEP|TIME", 
                   replacement = "", 
                   colnames(res$balance))
validParas 
### Flow mass balance components (inflow, outflow, storage):
vs2dh.plotMassBalance(data=res$balance)  
### Only resulting flow mass balance error (per timestep):
vs2dh.plotMassBalance(paras="FLUID__VOL_BAL", 
                       data = res$balance, 
                     mainLabel = "Flow mass balance error")  
### Energy mass balance components (inflow, outflow, storage):
vs2dh.plotMassBalance(paras=c("TOTAL__ENERGY_IN",
                              "TOTAL__ENERGY_OUT",
                              "ENERGY__STORAGE"),
                              data=res$balance,
                              mainLabel="Energy mass balance") 
### Only resulting energy mass balance error (per timestep):
vs2dh.plotMassBalance(paras="ENERGY__BALANCE",
                       data = res$balance,
                    mainLabel = "Energy mass balance error")   
}
