% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhPlot.R
\name{vs2dh.plotMatrix}
\alias{vs2dh.plotMatrix}
\title{Plot vs2dh.plotMatrix results}
\usage{
vs2dh.plotMatrix(data, nodes = TRUE, maxCol = 100,
  ignoreZeroValues = FALSE, ...)
}
\arguments{
\item{data}{matrix as returned by vs2di.run(model.path, returnOutput=TRUE)}

\item{nodes}{if TRUE node numbers are used as x-/y-axis; if false spatial extent}

\item{maxCol}{maximum number of colors to be used for colorramp}

\item{ignoreZeroValues}{if TRUE 0 values will be set to NA (Default: FALSE)}

\item{...}{further parameters passed to \code{levelplot()}
is derived returned in model units (currently not implemented!)}
}
\value{
Plot matrix values
}
\description{
Plot vs2dh.plotMatrix results
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)  
vs2dh.plotMatrix(data=res$variables$PressureHead[[1]])  
}
