% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhPlot.R
\name{vs2dh.plotObservationPoints}
\alias{vs2dh.plotObservationPoints}
\title{Spatial plotting of observation point time series}
\usage{
vs2dh.plotObservationPoints(paras = "Temp", data, paraLabel = NULL,
  maxCol = 100, ...)
}
\arguments{
\item{paras}{vector with at least one of the following parameters: H_m (total head),
P_m (pressure head), THETA (moisture content), SAT (saturation), TEMP 
(for temperature), VX (velocity in x-direction), VZ (velocity in z-direction), 
ET (evapotransporation)}

\item{data}{as returned by vs2di.run() in sublist "obsPoint"}

\item{paraLabel}{optional parameter for changing the label text of the parameters
defined in \code{paras} (if not used the value in "paras" is used for labelling)}

\item{maxCol}{maximum number of colors to be used for colorramp}

\item{...}{further parameters passed to \code{levelplot()}}
}
\value{
Spatial plots of observation point time series
}
\description{
Spatial plotting of observation point time series
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)  
vs2dh.plotObservationPoints (paras="TEMP",
                             data=res$obsPoints,
                             paraLabel="Temperature (\\u00B0\\ C)")  
}
