% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhPlot.R
\name{vs2dh.plotVariables}
\alias{vs2dh.plotVariables}
\title{Plot vs2dh.plotVariables matrix time series}
\usage{
vs2dh.plotVariables(para = "Temp", data, ignoreBoundary = TRUE,
  ignoreZeroValues = TRUE, paraLabel = NULL, maxCol = 100,
  main = "", nodes = TRUE, ...)
}
\arguments{
\item{para}{either "Temp" (for temperature) or "PressureHead"}

\item{data}{as returned by vs2di.run() in sublist "variables"}

\item{ignoreBoundary}{should boundary nodes (first/last column/row) be ignored?
Default: TRUE}

\item{ignoreZeroValues}{if TRUE 0 values will be set to NA (Default: TRUE)}

\item{paraLabel}{optional parameter for changing the label text of the parameter
(if not used the value in "para" is used for labelling)}

\item{maxCol}{maximum number of colors to be used for colorramp}

\item{main}{Additional text for main label (e.g. model name)}

\item{nodes}{if TRUE node numbers are used as x-/y-axis; if false spatial extent
is derived returned in model units (currently not implemented!)}

\item{...}{further parameters passed to \code{levelplot()}}
}
\value{
Time series matrix plot
}
\description{
Plot vs2dh.plotVariables matrix time series
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
modelName <- "tutorial2" 
model.path <- system.file("extdata", "vs2dh_example", modelName, package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)  
vs2dh.plotVariables ( para = "Temp",
                     data = res$variables,                    
                     paraLabel = "Temperature (\\u00B0\\ C)",
                     main = modelName)  
vs2dh.plotVariables (para = "PressureHead", 
                     data = res$variables, 
                     paraLabel = "Pressure head (m)",
                     main = modelName
                    )  
}
