% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhReadConfig.R
\name{vs2dh.ReadConfig}
\alias{vs2dh.ReadConfig}
\title{vs2dh.ReadConfig()}
\usage{
vs2dh.ReadConfig(model.path, dbg = TRUE)
}
\arguments{
\item{model.path}{Full path to the folder containing the vs2dh.dat}

\item{dbg}{prints debug information on the screen}
}
\value{
Imported vs2dh configuration
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
vs2dh.ReadConfig(model.path)
}
\description{
Imports the configuration from an existing vs2dh.dat file
and stores it in an R list data structure as retrieved by vs2dh.Configure()
}

