% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhReadOutput.R
\name{vs2dh.readObsPoints}
\alias{vs2dh.readObsPoints}
\title{Read VS2dh model output file with observation point time series}
\usage{
vs2dh.readObsPoints(model.path, fileName = "obsPoints.out", dbg = TRUE)
}
\arguments{
\item{model.path}{full path to folder containing model output files}

\item{fileName}{name of file containing observation point results 
(default: "obsPoints.out")}

\item{dbg}{if true text output}
}
\value{
read model results at observation points aare imported in a R object
}
\description{
Read VS2dh model output file with observation point time series
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
vs2dh.readObsPoints(model.path)
}
