% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vs2dhReadOutput.R
\name{vs2dh.readVariables}
\alias{vs2dh.readVariables}
\title{Read VS2dh model output file with time series of variables pressure head and
temperature}
\usage{
vs2dh.readVariables(model.path, fileName = "variables.out", dbg = TRUE)
}
\arguments{
\item{model.path}{full path to folder containing model output files}

\item{fileName}{name of file containing variable results
(default: "variables.out")}

\item{dbg}{if true text output}
}
\value{
variable model results are imported in a R object
}
\description{
Read VS2dh model output file with time series of variables pressure head and
temperature
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)
vs2dh.readVariables(model.path)
}

