% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhWriteConfig.R
\name{vs2dh.writeConfig}
\alias{vs2dh.writeConfig}
\title{Write configuration}
\usage{
vs2dh.writeConfig(conf)
}
\arguments{
\item{conf}{model config as retrieved by vs2dhConfigure()}
}
\value{
Write vs2dh.dat in folder \code{model.path}
}
\description{
Write configuration
}
\examples{
conf <- vs2dh.Configure()
inpDat <- vs2dh.writeConfig(conf)  
write(inpDat, file.path(getwd(), "inp.dat"))
}
