% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhRunModel.R
\name{vs2di.createFileDat}
\alias{vs2di.createFileDat}
\title{Create vs2dh.fil or vs2dt.fil file}
\usage{
vs2di.createFileDat(engine, model.path, dbg = TRUE)
}
\arguments{
\item{engine}{model engine either 'vs2dh' or 'vs2dt'}

\item{model.path}{full path to folder containing vs2dh.dat}

\item{dbg}{if true text output}
}
\value{
Saves/Overwrites vs2dh.fil in folder \code{model.path}
}
\description{
Create vs2dh.fil or vs2dt.fil file
}
\examples{
### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
vs2di.createFileDat(engine = "vs2dh", 
                    model.path = model.path)
}
