% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2dhRunModel.R
\name{vs2di.run}
\alias{vs2di.run}
\title{Run VS2dh model}
\usage{
vs2di.run(engine = "vs2dh",
  engineDirectoryWin = system.file("extdata/engine/win", package =
  "kwb.vs2dh"), engineDirectoryLinux = system.file("extdata/engine/linux",
  package = "kwb.vs2dh"), model.path = system.file("extdata",
  "vs2dh_example/tutorial2", package = "kwb.vs2dh"), returnOutput = TRUE,
  showWarnings = TRUE, openTargetDir = FALSE, dbg = TRUE)
}
\arguments{
\item{engine}{model engine 'vs2dh' (for flow & heat modelling) or 'vs2dt' (for 
flow & solute transport) (Default: "vs2dh")}

\item{engineDirectoryWin}{default directory on Windows OS containing vs2dh3_3.exe and 
vs2dt3_3.exe (Default: system.file("extdata/engine/win", package = "kwb.vs2dh"))}

\item{engineDirectoryLinux}{default directory on Linux OS containing vs2dh3_3.exe and 
vs2dt3_3.exe (Default: system.file("extdata/engine/linux", package = "kwb.vs2dh"))}

\item{model.path}{full path to folder containing vs2dh.dat (default: 
system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh"))}

\item{returnOutput}{if TRUE model output files variables.out, balance.out and
obsPoints.out are imported into R completely; vs2dh.out (only warnings imported)}

\item{showWarnings}{if TRUE print warning messages during simulation on screen 
(default: TRUE)}

\item{openTargetDir}{If TRUE path containing model files will be opened in 
explorer (Default: FALSE)}

\item{dbg}{if true text output on screen on model run progress}
}
\value{
Run VS2dh model and saves model output files in folder \code{model.path}.
If \code{returnOutput} is TRUE the output is imported in R object
}
\description{
Run VS2dh model
}
\examples{
### Running model with default model.path and engine.path: 
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)
}
